Split of ExplainEditorJS
In this commit we split ExplainEditorJS.scala into 4 files
Ajaxer.scala : autowire implementation and Model.
ExplainEditorJSDomBuilders.scala : containing dom generating code
ExplainEditorPresenceHelpers.scala: for presence wrappers
ExplainEditorJS.scala: Essentially code visible from plain JavaScript
